feat(data-parser): added blocks and txs types
